requirements.txt
adal==1.2.4
allure-behave==2.8.18
allure-python-commons==2.8.18
Appium-Python-Client==1.0.2
attrs==20.2.0
behave==1.2.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
chromedriver-autoinstaller==0.2.2
cryptography==3.1.1
decorator==4.4.2
defusedxml==0.6.0
docutils==0.16
idna==2.10
importlib-metadata==2.0.0
jira==2.0.0
kitchen==1.2.6
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.3
parse==1.18.0
parse-type==0.5.2
pbr==5.5.0
pluggy==0.13.1
pycparser==2.20
PyJWT==1.7.1
pyodbc==4.0.30
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
selenium==3.141.0
six==1.15.0
urllib3==1.25.10
xlrd==1.2.0
zipp==3.2.0
